requirements.txt
Django==3.2.7
python-telegram-bot==13.7
psycopg2==2.9.1
python-decouple==3.4
text-unidecode==1.3
django-mptt==0.12.0
